🏷️rename: MainCard => TemplateItem
