+++ src/components/server/TemplateItem/TemplateItem.type.ts
@@ -1,4 +1,4 @@
-export interface MainCardProps {
+export interface TemplateItemProps {
   children: React.ReactNode;
   name: string;
   price: number | string;
