+++ src/components/server/index.ts
@@ -11,4 +11,4 @@ export { default as CommonList } from './CommonList/CommonList';
 export { default as Banner } from './Banner/Banner';
 export { default as TemplateCard } from './TemplateCard/TemplateCard';
 export { default as PhoneMockup } from './PhoneMockUp/PhoneMockup';
-export { default as MainCard } from './MainCard/MainCard';
+export { default as TemplateItem } from './TemplateItem/TemplateItem';
